PHPStan L8 全量清零：279 → 0，主配置 level 7 → 8
错误消解（279 → 0）：
- Controller 基类新增 protected user(): User（未认证 abort 401，与
  auth 中间件语义一致），48 个控制器 185 处 $request->user() /
  Auth::user() 统一收口为 $this->user()，静态分析随 helper 返回
  类型自动窄化
- Api/v1 路由 action user() 与基类 helper 冲突，改名 me()（路由
  同步更新）；AnalyticsController / WebsiteDataController /
  ResourcesController / DomainController / WebsiteController 补齐
  基类继承
- SeoAuditController 游客路径（analyze / accessState）保留 nullable
  的 auth()->user()，不强制 401，游客限流语义不变
- SeoCheckTools::fetchPage 增加 union array-shape @return，phpstan
  按 ok 分支窄化 dom；18 处 $page['error'] ?? 兜底随 shape 化去冗余
- nullable 传参收口：preg_replace/first/find/关系属性等 null 链路
  逐点 (string) 强转、?-> 窄化或显式 null 检查（登录密码空值、
  离线支付缺关联用户 fail-loud、PlanLimit 配额降级不限、团队
  无归属拒绝等）
- 文档类型修正：DomainMonitor::whois / StatisticsService::
  breakdownWithUtm 的 @return shape 与实际对齐
- 回归测试断言同步：Tasks25RegressionTest 源码断言跟随
  $this->user() 写法

回归验证：
- pint 全仓 489 文件 PASS
- 608 tests / 2328 assertions 全绿
- HTTP 冒烟 landing / login 均 200
- phpstan.neon level 8 + 空 baseline 本地全量复验通过

+++ app/Http/Controllers/IndexController.php
@@ -52,8 +52,8 @@ class IndexController extends Controller
         // 定价卡：优先 prices 直配价，无则按默认货币价 × 汇率换算（规格书 §10.4）
         // 同时输出年付价（原站定价区月/年切换 + 折扣徽章）
         $plans = Plan::where('is_enabled', true)->orderBy('order')->get()->map(function (Plan $plan) use ($currency) {
-            $plan->landing_price = Currency::planPrice($plan, $currency, 'monthly');
-            $plan->landing_price_annual = Currency::planPrice($plan, $currency, 'annual');
+            $plan->landing_price = (float) Currency::planPrice($plan, $currency, 'monthly');
+            $plan->landing_price_annual = (float) Currency::planPrice($plan, $currency, 'annual');
             $plan->landing_currency = $currency;
 
             return $plan;
